+++ packages/typespec-rust/CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed an issue that could cause types to have duplicate names.
 * Fixed an issue that could cause enum types to have invalid names.
 * Ensure that the local variable name for the `http::Request` doesn't collide with an existing parameter name.
+* Added missing reinjectable query parameters when creating the `Url` for a pageable's `next_link`.
 
 ### Other Changes